Add support for loop folding in detector error modelling mode (#47)
- Refactor `REPEAT` block to allow repetition counts up to 2^63-1.
- Fix each `OBSERVABLE_INCLUDE` instruction counting towards estimated memory usage in `--detect` mode (instead of the max observable index that was used)
- Increased supported sample counts up to `INT64_MAX`
- Fix using strtol when parsing command line args, because it silently accepts too-large inputs while incorrectly parsing them
- Refactor `find_int_argument` into `find_int64_argument`
- Add tests of exact out-of-bounds limits of `find_int64_argument`
- Change `CircuitGenParameters.rounds` to a `uint64_t`
- Parse shots as `uint64_t` in `--sample` and `--detect` modes
- Add `--fold_loops` command line option
- Increase detector id size from uint32_t to uint64_t (since loops can massively expand them)
- Add `accumulate_errors` field to `ErrorFuser`, to disable error operations
- Add `fold_loops` field to `ErrorFuser`, to enable loop folding
- Add `ErrorFuser::run_loops` method that uses tortoise+hare cycle finding
- Refactor ErrorFuser to build a `FusedError` vector that then gets printed (instead of directly printing)
- Add `ErrorFuser::flush` for flushing the error probability map into the fused error vector
- Add `ErrorFuser::shift_active_detector_ids` for applying many iterations of a loop when known to be at a periodic fixed point

+++ src/simulators/frame_simulator.cc
@@ -492,7 +492,7 @@ void sample_out_helper(
 }
 
 void FrameSimulator::sample_out(
-    const Circuit &circuit, const simd_bits &reference_sample, size_t num_shots, FILE *out, SampleFormat format,
+    const Circuit &circuit, const simd_bits &reference_sample, uint64_t num_shots, FILE *out, SampleFormat format,
     std::mt19937_64 &rng) {
     constexpr size_t GOOD_BLOCK_SIZE = 1024;
     size_t num_qubits = circuit.count_qubits();
